Buggy and untested music support for other sources
