+++ src/commands/music/package.json
@@ -1,7 +1,8 @@
 {
   "main": "music.js",
   "dependencies": {
-    "ytdl-core": "^0.9.0",
-    "youtube-search": "^1.0.8"
+    "youtube-dl": "^1.11.1",
+    "youtube-search": "^1.0.8",
+    "request": "^2.79.0"
   }
 }
